remove the bench
